+++ automne/classes/modules/modulescatalog.php
@@ -14,7 +14,7 @@
 // | Author: Sébastien Pauchet <sebastien.pauchet@ws-interactive.fr>      |
 // +----------------------------------------------------------------------+
 //
-// $Id: modulescatalog.php,v 1.2 2009/06/05 15:02:19 sebastien Exp $
+// $Id: modulescatalog.php,v 1.3 2009/10/22 16:30:02 sebastien Exp $
 
 /**
   * Class CMS_modulesCatalog
@@ -235,7 +235,7 @@ class CMS_modulesCatalog extends CMS_grandFather
 		$tables_prefixes = array();
 		while ($data = $q->getArray()) {
 			if (ereg($tablesPrefix."(.*)_public", $data[0])) { //TODOV4
-				$tables_prefixes[] = substr($data[0], 0, strrpos($data[0], "_") + 1);
+				$tables_prefixes[] = io::substr($data[0], 0, strrpos($data[0], "_") + 1);
 			}
 		}
 		
@@ -289,37 +289,46 @@ class CMS_modulesCatalog extends CMS_grandFather
 				return false;
 			}
 			//delete all files of the locationToDir
-			foreach(glob(PATH_MODULES_FILES_FS."/".$module->getCodename()."/".$locationTo.'/r'.$resourceID.'_*', GLOB_NOSORT) as $file) {
-				if (!CMS_file::deleteFile($file)) {
-					CMS_grandFather::raiseError("Can't delete file ".$file);
-					return false;
+			$files = glob(PATH_MODULES_FILES_FS."/".$module->getCodename()."/".$locationTo.'/r'.$resourceID.'_*', GLOB_NOSORT);
+			if (is_array($files)) {
+				foreach($files as $file) {
+					if (!CMS_file::deleteFile($file)) {
+						CMS_grandFather::raiseError("Can't delete file ".$file);
+						return false;
+					}
 				}
 			}
 			//then copy or move them to the locationToDir
-			foreach(glob(PATH_MODULES_FILES_FS."/".$module->getCodename()."/".$locationFrom.'/r'.$resourceID.'_*', GLOB_NOSORT) as $file) {
-				$to = str_replace('/'.$locationFrom.'/','/'.$locationTo.'/',$file);
-				if ($copyOnly) {
-					if (!CMS_file::copyTo($file,$to)) {
-						CMS_grandFather::raiseError("Can't copy file ".$file." to ".$to);
-						return false;
+			$files = glob(PATH_MODULES_FILES_FS."/".$module->getCodename()."/".$locationFrom.'/r'.$resourceID.'_*', GLOB_NOSORT);
+			if (is_array($files)) {
+				foreach($files as $file) {
+					$to = str_replace('/'.$locationFrom.'/','/'.$locationTo.'/',$file);
+					if ($copyOnly) {
+						if (!CMS_file::copyTo($file,$to)) {
+							CMS_grandFather::raiseError("Can't copy file ".$file." to ".$to);
+							return false;
+						}
+					} else {
+						if (!CMS_file::moveTo($file,$to)) {
+							CMS_grandFather::raiseError("Can't move file ".$file." to ".$to);
+							return false;
+						}
 					}
-				} else {
-					if (!CMS_file::moveTo($file,$to)) {
-						CMS_grandFather::raiseError("Can't move file ".$file." to ".$to);
-						return false;
-					}
+					//then chmod new file
+					CMS_file::chmodFile(FILES_CHMOD,$to);
 				}
-				//then chmod new file
-				CMS_file::chmodFile(FILES_CHMOD,$to);
 			}
 		}
 		//cleans the initial dir if not a copy
 		if (!$copyOnly) {
 			//then get all files of the locationFromDir
-			foreach(glob(PATH_MODULES_FILES_FS."/".$module->getCodename()."/".$locationFrom.'/r'.$resourceID.'_*', GLOB_NOSORT) as $file) {
-				if (!CMS_file::deleteFile($file)) {
-					CMS_grandFather::raiseError("Can't delete file ".$file);
-					return false;
+			$files = glob(PATH_MODULES_FILES_FS."/".$module->getCodename()."/".$locationFrom.'/r'.$resourceID.'_*', GLOB_NOSORT);
+			if (is_array($files)) {
+				foreach($files as $file) {
+					if (!CMS_file::deleteFile($file)) {
+						CMS_grandFather::raiseError("Can't delete file ".$file);
+						return false;
+					}
 				}
 			}
 		}
@@ -343,7 +352,7 @@ class CMS_modulesCatalog extends CMS_grandFather
 			$destination_dir = dir($destination_path);
 			while (false !== ($file = $destination_dir->read())) {
 				if (is_file($destination_path."/".$file)
-					&& substr($file, 0, strlen($files_prefix)) == $files_prefix) {
+					&& io::substr($file, 0, io::strlen($files_prefix)) == $files_prefix) {
 					unlink($destination_path."/".$file);
 				}
 			}
@@ -351,7 +360,7 @@ class CMS_modulesCatalog extends CMS_grandFather
 			//copy or move the files
 			while (false !== ($file = $initial_dir->read())) {
 				if (is_file($initial_path."/".$file)
-					&& substr($file, 0, strlen($files_prefix)) == $files_prefix) {
+					&& io::substr($file, 0, io::strlen($files_prefix)) == $files_prefix) {
 					if ($copyOnly) {
 						if (@copy($initial_path."/".$file, $destination_path."/".$file)) {
 							@chmod($destination_path."/".$file, octdec(FILES_CHMOD));
@@ -374,7 +383,7 @@ class CMS_modulesCatalog extends CMS_grandFather
 			$initial_dir->rewind();
 			while (false !== ($file = $initial_dir->read())) {
 				if (is_file($initial_path."/".$file)
-					&& substr($file, 0, strlen($files_prefix)) == $files_prefix) {
+					&& io::substr($file, 0, io::strlen($files_prefix)) == $files_prefix) {
 					unlink($initial_path."/".$file);
 				}
 			}
